+++ elfo-core/src/context.rs
@@ -22,7 +22,7 @@ use crate::{
     messages,
     request_table::ResponseToken,
     routers::Singleton,
-    scope,
+    scope, trace_id,
 };
 
 pub(crate) use self::source::Source;
@@ -45,7 +45,7 @@ pub struct Context<C = (), K = Singleton, S = ()> {
 
 assert_impl_all!(Context: Send);
 
-impl<C, K, S> Context<C, K, S> {
+impl<C: 'static, K, S> Context<C, K, S> {
     #[inline]
     pub fn addr(&self) -> Addr {
         self.addr
@@ -91,6 +91,15 @@ impl<C, K, S> Context<C, K, S> {
         let object = ward!(self.book.get_owned(self.addr));
         let actor = ward!(object.as_actor());
         actor.set_status(status);
+    }
+
+    /// Closes the mailbox, that leads to returning `None` from `recv()` and
+    /// `try_recv()` after handling all available messages in the mailbox.
+    ///
+    /// Returns `true` if the mailbox has just been closed.
+    pub fn close(&self) -> bool {
+        let object = ward!(self.book.get_owned(self.addr), return false);
+        ward!(object.as_actor(), return false).close()
     }
 
     pub async fn send<M: Message>(&self, message: M) -> Result<(), SendError<M>> {
@@ -235,7 +244,6 @@ impl<C, K, S> Context<C, K, S> {
 
     pub async fn recv(&mut self) -> Option<Envelope>
     where
-        C: 'static,
         S: Source,
     {
         self.stats.message_handling_time_seconds();
@@ -244,6 +252,10 @@ impl<C, K, S> Context<C, K, S> {
         let object = self.book.get_owned(self.addr)?;
         let actor = object.as_actor()?;
 
+        if actor.is_initializing() {
+            actor.set_status(ActorStatus::NORMAL);
+        }
+
         // TODO: remove `fuse`.
         let mailbox_fut = actor.recv().fuse();
         let source_fut = poll_fn(|cx| self.source.poll_recv(cx)).fuse();
@@ -251,57 +263,52 @@ impl<C, K, S> Context<C, K, S> {
         pin_mut!(mailbox_fut);
         pin_mut!(source_fut);
 
-        // TODO: reset trace_id for these logs?
         let envelope = select_biased! {
-            envelope = mailbox_fut => match envelope {
-                Some(envelope) => envelope,
-                None => {
-                    trace!("mailbox closed");
-                    return None;
-                }
-            },
-            envelope = source_fut => match envelope {
-                Some(envelope) => envelope,
-                None => {
-                    // TODO: rerun select?
-                    trace!("some of sources was closed");
-                    return None;
-                }
-            },
+            envelope = mailbox_fut => envelope,
+            envelope = source_fut => envelope, // TODO: rerun select if `None`?
         };
 
-        Some(self.post_recv(envelope))
+        match envelope {
+            Some(envelope) => Some(self.post_recv(envelope)),
+            None => {
+                actor.set_status(ActorStatus::TERMINATING);
+                // TODO: forward Terminate's trace_id.
+                scope::set_trace_id(trace_id::generate());
+                trace!("input closed");
+                None
+            }
+        }
     }
 
-    pub fn try_recv(&mut self) -> Result<Envelope, TryRecvError>
-    where
-        C: 'static,
-    {
+    pub fn try_recv(&mut self) -> Result<Envelope, TryRecvError> {
         self.stats.message_handling_time_seconds();
 
         let object = self.book.get(self.addr).ok_or(TryRecvError::Closed)?;
         let actor = object.as_actor().ok_or(TryRecvError::Closed)?;
-        let envelope = match actor.try_recv() {
-            Ok(envelope) => envelope,
-            Err(err) => {
-                if err.is_closed() {
-                    // TODO: reset trace_id for this log?
-                    trace!("mailbox closed");
-                }
-                return Err(err);
-            }
-        };
-        drop(object);
+
+        if actor.is_initializing() {
+            actor.set_status(ActorStatus::NORMAL);
+        }
 
         // TODO: poll the sources.
-
-        Ok(self.post_recv(envelope))
+        match actor.try_recv() {
+            Ok(envelope) => {
+                drop(object);
+                Ok(self.post_recv(envelope))
+            }
+            Err(err) => {
+                if err.is_closed() {
+                    actor.set_status(ActorStatus::TERMINATING);
+                    // TODO: forward Terminate's trace_id.
+                    scope::set_trace_id(trace_id::generate());
+                    trace!("mailbox closed");
+                }
+                Err(err)
+            }
+        }
     }
 
-    fn post_recv(&mut self, envelope: Envelope) -> Envelope
-    where
-        C: 'static,
-    {
+    fn post_recv(&mut self, envelope: Envelope) -> Envelope {
         scope::set_trace_id(envelope.trace_id());
 
         let envelope = msg!(match envelope {
@@ -314,7 +321,12 @@ impl<C, K, S> Context<C, K, S> {
                 self.respond(token, Ok(()));
                 envelope
             }
-            envelope => envelope,
+            envelope => {
+                if envelope.is::<messages::Terminate>() {
+                    self.set_status(ActorStatus::TERMINATING);
+                }
+                envelope
+            }
         });
 
         let message = envelope.message();
@@ -332,14 +344,19 @@ impl<C, K, S> Context<C, K, S> {
         }
 
         self.stats.message_waiting_time_seconds(&envelope);
-
         envelope
+    }
+
+    /// This is a part of private API for now.
+    #[doc(hidden)]
+    pub async fn finished(&self, addr: Addr) {
+        ward!(self.book.get_owned(addr)).finished().await;
     }
 
     /// XXX: mb `BoundEnvelope<C>`?
     pub fn unpack_config<'c>(&self, config: &'c AnyConfig) -> &'c C
     where
-        C: for<'de> serde::Deserialize<'de> + 'static,
+        C: for<'de> serde::Deserialize<'de>,
     {
         config.get_user()
     }